pyproject.toml
[tool.poetry]
name = "solid_state"
version = "0.1.0"
description = ""
authors = ["Ben Baker-Smith <bbakersmith@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
parsimonious = "^0.9.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
solidpython = "^1.1.3"
black = "^22.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
